linux uses 64bit for pointers

+++ src/graphics/sector_queue/sector_queue_push.c
@@ -6,7 +6,7 @@
 /*   By: mfischer <mfischer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/03 14:45:09 by mfischer          #+#    #+#             */
-/*   Updated: 2019/09/01 17:42:49 by mfischer         ###   ########.fr       */
+/*   Updated: 2019/09/01 20:56:37 by mfischer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ t_sector	*sector_queue_push(t_sector *sector)
 	if (sector == NULL)
 		return (stack_pop(sector_queue));
 	stack_push(sector_queue, sector);
-	if ((int32_t)sector == -1)
+	if ((int64_t)sector == -1)
 		stack_destroy(&sector_queue);
 	return (NULL);
 }
